remove parameters packages in rao-commons
