+++ ra-optimisation/rao-commons/src/main/java/com/farao_community/farao/rao_commons/systematic_sensitivity/SystematicSensitivityComputation.java
@@ -13,12 +13,10 @@ import com.farao_community.farao.data.crac_result_extensions.CnecResult;
 import com.farao_community.farao.data.crac_result_extensions.CnecResultExtension;
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.farao_community.farao.rao_commons.RaoData;
-import com.farao_community.farao.rao_commons.systematic_sensitivity.parameters.SystematicSensitivityComputationParameters;
 import com.farao_community.farao.util.SensitivityComputationException;
 import com.farao_community.farao.util.SystematicSensitivityAnalysisResult;
 import com.farao_community.farao.util.SystematicSensitivityAnalysisService;
 import com.powsybl.computation.ComputationManager;
-import com.powsybl.computation.DefaultComputationManagerConfig;
 import com.powsybl.sensitivity.SensitivityComputationParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -58,21 +56,6 @@ public class SystematicSensitivityComputation {
     /**
      * Constructors
      */
-    public SystematicSensitivityComputation() {
-        this(new SystematicSensitivityComputationParameters(),
-            DefaultComputationManagerConfig.load().createShortTimeExecutionComputationManager());
-    }
-
-    public SystematicSensitivityComputation(RaoParameters raoParameters) {
-        SystematicSensitivityComputationParameters parameters;
-        if (!Objects.isNull(raoParameters.getExtension(SystematicSensitivityComputationParameters.class))) {
-            parameters = raoParameters.getExtension(SystematicSensitivityComputationParameters.class);
-        } else {
-            parameters = new SystematicSensitivityComputationParameters();
-        }
-        init(parameters);
-    }
-
     public SystematicSensitivityComputation(RaoParameters raoParameters, ComputationManager computationManager) {
         SystematicSensitivityComputationParameters parameters;
         if (!Objects.isNull(raoParameters.getExtension(SystematicSensitivityComputationParameters.class))) {
@@ -83,16 +66,8 @@ public class SystematicSensitivityComputation {
         init(parameters, computationManager);
     }
 
-    public SystematicSensitivityComputation(SystematicSensitivityComputationParameters computationParameters) {
-        init(computationParameters);
-    }
-
     public SystematicSensitivityComputation(SystematicSensitivityComputationParameters computationParameters, ComputationManager computationManager) {
         init(computationParameters, computationManager);
-    }
-
-    private void init(SystematicSensitivityComputationParameters computationParameters) {
-        init(computationParameters, DefaultComputationManagerConfig.load().createShortTimeExecutionComputationManager());
     }
 
     private void init(SystematicSensitivityComputationParameters computationParameters, ComputationManager computationManager) {
