+++ ra-optimisation/rao-commons/src/main/java/com/farao_community/farao/rao_commons/systematic_sensitivity/SystematicSensitivityComputationParameters.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.farao_community.farao.rao_commons.systematic_sensitivity.parameters;
+package com.farao_community.farao.rao_commons.systematic_sensitivity;
 
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.powsybl.commons.extensions.AbstractExtension;
