+++ ra-optimisation/rao-commons/src/test/java/com/farao_community/farao/rao_commons/linear_optimisation/core/JsonLinearProblemParametersTest.java
@@ -10,7 +10,6 @@ package com.farao_community.farao.rao_commons.linear_optimisation.core;
 import com.farao_community.farao.commons.FaraoException;
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.farao_community.farao.rao_api.json.JsonRaoParameters;
-import com.farao_community.farao.rao_commons.linear_optimisation.core.parameters.LinearProblemParameters;
 import com.powsybl.commons.AbstractConverterTest;
 import org.junit.Test;
