+++ ra-optimisation/rao-commons/src/test/java/com/farao_community/farao/rao_commons/systematic_sensitivity/SystematicSensitivityComputationConfigLoaderTest.java
@@ -7,8 +7,6 @@
 
 package com.farao_community.farao.rao_commons.systematic_sensitivity;
 
-import com.farao_community.farao.rao_commons.systematic_sensitivity.parameters.SystematicSensitivityComputationConfigLoader;
-import com.farao_community.farao.rao_commons.systematic_sensitivity.parameters.SystematicSensitivityComputationParameters;
 import com.powsybl.commons.config.PlatformConfig;
 import com.powsybl.sensitivity.SensitivityComputationParameters;
 import org.junit.Before;
